move es utils into nested folder
