+++ seqr/views/status.py
@@ -4,7 +4,7 @@ import redis
 from urllib3.connectionpool import connection_from_url
 
 from settings import SEQR_VERSION, KIBANA_SERVER, REDIS_SERVICE_HOSTNAME, REDIS_SERVICE_PORT, DATABASES
-from seqr.utils.elasticsearch.utils import get_es_client
+from seqr.utils.search.utils import get_es_client
 from seqr.views.utils.json_utils import create_json_response
 
 logger = logging.getLogger(__name__)
